Add graphviz runtime dep for API reference
Given the sphinx configuration provided by documenteer, graphviz is
needed when building documentation for Science Pipelines because it is
called by automodapi to generate dependency diagrams. Thus this
dependency in the lsst-documenteer-pipelines metapackage ensures
graphviz is available. Note that graphviz is not included as a
dependency in the automodapi conda-forge recipe.