Drop lsst-documenteer-technote package
Because we now build technotes in tox/nox virtual environments, there's not a real demand for having documenteer[technote] in the environment. As well, we'd have to start packaging `technote` into conda-forge, which I'm reluctant to do. If this causes issues we can re-evaluate, but generally pushing technote users away from conda wherever possible is a great thing for me and my user support burden.